chore: addresses sonarcloud issues, removes unused code
